candle_nn

Trait ModuleT

pub trait ModuleT {
    // Required method
    fn forward_t(&self, xs: &Tensor, train: bool) -> Result<Tensor, Error>;
}

A single forward method using a single single tensor argument and a flag to separate the training and evaluation behaviors.

Required Methods

fn forward_t(&self, xs: &Tensor, train: bool) -> Result<Tensor, Error>